requirements.txt
holehe
python-dotenv